trustlayer-ai/requirements.txt
numpy>=1.26.0
scipy>=1.14.0
pillow>=10.0.0
matplotlib>=3.9.0
pydub  # Optional - needs ffmpeg for MP3